introduce build qualifiers
the 3rd member of the build tuple is now a keyword list, where options
can live that specify a target to a finer degree.

For example:
`{:os, :cpu, libc: ...}` can be used to specify the libc variant in a target.

Built targets are also now aliases with names, which are used by
the `BURRITO_TARGET` env variable to override what target is being selected.

We still support the old build target strings: [:linux, :linux_musl, :darwin, :win32]
and they will be converted internally to the new target structure.

**This support will be removed in future releases**